+++ src/lib.rs
@@ -1,6 +1,6 @@
 //! Track and settle debts between parties with the fewest transactions.
 //!
-//! Debtsolves gives you two structs - Transactions, which track payments that
+//! Debtsolver gives you two structs - Transactions, which track payments that
 //! have been made or need to be made, and a Ledger which can store and balance
 //! the current states of credits and debits between everyone.
 //!
@@ -77,31 +77,43 @@
 //!     // Bob owes Alice 15
 //!     // Bob owes Charlie 15
 //! ```
-
 use itertools::Itertools;
 use std::cmp;
 use std::collections::HashMap;
 use std::error::Error;
 use std::fmt;
+use rusty_money::money; 
+use rusty_money::Currency;
+use rusty_money::Money;
+
+// Branch TODOs
+// Audit all to-do's
+// Strict error and panic assertions (match type and message)
 
 /// Represents a transaction where one party (debtor) pays another (creditor) the amount specified.
 #[derive(Debug, Eq, Ord, PartialEq, PartialOrd)]
 pub struct Transaction {
     debtor: String,
     creditor: String,
-    amount: i32,
+    amount: Money,
+}
+
+macro_rules! transaction {
+    ($x:expr, $y:expr, $z:expr) => {
+        Transaction::new($x.to_string(), $y.to_string(), $z).unwrap()
+    };
 }
 
 impl Transaction {
-    pub fn new(debtor: String, creditor: String, amount: i32) -> Result<Self, ParseAmountError> {
-        if amount <= 0 {
-            return Err(ParseAmountError { amount: amount });
+    pub fn new(debtor: String, creditor: String, amount: Money) -> Result<Self, ParseAmountError> {
+        if !amount.is_positive() {
+            return Err(ParseAmountError { amount });
         };
-        return Ok(Transaction {
+        Ok(Transaction {
             debtor,
             creditor,
             amount,
-        });
+        })
     }
 }
 
@@ -117,23 +129,23 @@ impl fmt::Display for Transaction {
 pub struct MultiPartyTransaction {
     debtors: Vec<String>,
     creditors: Vec<String>,
-    amount: i32,
+    amount: Money,
 }
 
 impl MultiPartyTransaction {
     pub fn new(
         debtors: Vec<String>,
         creditors: Vec<String>,
-        amount: i32,
+        amount: Money,
     ) -> Result<Self, ParseAmountError> {
-        if amount <= 0 {
-            return Err(ParseAmountError { amount: amount });
+        if amount.is_negative() {
+            return Err(ParseAmountError { amount });
         };
-        return Ok(MultiPartyTransaction {
+        Ok(MultiPartyTransaction {
             debtors,
             creditors,
             amount,
-        });
+        })
     }
 }
 
@@ -151,7 +163,7 @@ impl fmt::Display for MultiPartyTransaction {
 
 #[derive(Debug)]
 pub struct ParseAmountError {
-    amount: i32,
+    amount: Money,
 }
 
 impl Error for ParseAmountError {}
@@ -170,48 +182,42 @@ impl fmt::Display for ParseAmountError {
 /// The sum of all balances must always add up to zero, since each debtor has an equivalent creditor.
 #[derive(Debug)]
 pub struct Ledger {
-    map: HashMap<String, i32>,
+    map: HashMap<String, Money>,
 }
 
 impl Ledger {
     /// Creates a new Ledger
     pub fn new() -> Ledger {
-        return Ledger {
+        Ledger {
             map: HashMap::new(),
-        };
+        }
     }
 
     /// Accepts a transaction and updates debtor and creditor balances in the ledger.
     pub fn add_transaction(&mut self, transaction: Transaction) {
-        *self.map.entry(transaction.debtor).or_insert(0) -= transaction.amount;
-        *self.map.entry(transaction.creditor).or_insert(0) += transaction.amount;
+        *self
+            .map
+            .entry(transaction.debtor)
+            .or_insert_with(|| money!(0, "USD")) -= transaction.amount.clone();
+        *self
+            .map
+            .entry(transaction.creditor)
+            .or_insert_with(|| money!(0, "USD")) += transaction.amount.clone();
     }
 
     pub fn add_multi_party_transaction(&mut self, transaction: MultiPartyTransaction) {
         let num_debtors = transaction.debtors.len() as i32;
-        let debt_amount = transaction.amount / num_debtors;
-        let debt_total = num_debtors * debt_amount;
+        let mut debt_shares = transaction.amount.allocate_to(num_debtors);
+        for debtor in transaction.debtors {
+            *self.map.entry(debtor).or_insert_with(|| money!(0, "USD")) -=
+                debt_shares.pop().unwrap();
+        }
 
         let num_creditors = transaction.creditors.len() as i32;
-        let credit_amount = transaction.amount / num_creditors;
-        let credit_total = num_creditors * credit_amount;
-
-        if debt_total > credit_total {
-            *self
-                .map
-                .entry(transaction.creditors[0].clone())
-                .or_insert(0) += debt_total - credit_total;
-        } else if credit_total > debt_total {
-            *self.map.entry(transaction.debtors[0].clone()).or_insert(0) -=
-                credit_total - debt_total;
-        }
-
-        for debtor in transaction.debtors {
-            *self.map.entry(debtor).or_insert(0) -= debt_amount;
-        }
-
+        let mut credit_shares = transaction.amount.allocate_to(num_creditors);
         for creditor in transaction.creditors {
-            *self.map.entry(creditor).or_insert(0) += credit_amount;
+            *self.map.entry(creditor).or_insert_with(|| money!(0, "USD")) +=
+                credit_shares.pop().unwrap();
         }
     }
 
@@ -221,28 +227,28 @@ impl Ledger {
     pub fn settle(&mut self, group_size: usize) -> Vec<Transaction> {
         let mut payments: Vec<Transaction> = Vec::new();
         if group_size > 0 {
-            for x in 1..group_size + 1 {
+            for x in 1..=group_size {
                 payments.append(&mut self.settle_combinations(x));
             }
         }
         payments.append(&mut self.clear_all_entries());
-        return payments;
+        payments
     }
 
     // Converts the ledger from a hashmap into a set of vector-tuples containing the
     // debtor/creditor and the amount. Debts are negative, and credits are positive.
-    pub fn to_vector(&self) -> Vec<(String, i32)> {
-        let mut ledger_entries: Vec<(String, i32)> = Vec::new();
+    pub fn to_vector(&self) -> Vec<(String, Money)> {
+        let mut ledger_entries: Vec<(String, Money)> = Vec::new();
 
         for (key, val) in self.map.iter() {
-            ledger_entries.push((key.clone(), *val));
+            ledger_entries.push((key.clone(), val.clone()));
         }
-        return ledger_entries;
+        ledger_entries
     }
 
     fn panic_unless_empty(&self) {
         for (_, val) in self.map.iter() {
-            if *val > 0 {
+            if !val.is_zero() {
                 panic!();
             }
         }
@@ -259,16 +265,16 @@ impl Ledger {
             let mut debtor_keys: Vec<String> = Vec::new();
             let mut creditor_keys: Vec<String> = Vec::new();
             for item in combo {
-                if item.1 > 0 {
+                if item.1.is_positive() {
                     creditor_keys.push(item.0)
-                } else if item.1 < 0 {
+                } else if item.1.is_negative() {
                     debtor_keys.push(item.0)
                 } else {
                 }
             }
             payments.append(&mut self.clear_given_keys(debtor_keys, creditor_keys));
         }
-        return payments;
+        payments
     }
 
     // Settles all entries left in the ledger with a balance, in random order.
@@ -276,10 +282,10 @@ impl Ledger {
         let (debtor_keys, creditor_keys) = self.debtor_and_creditor_keys();
         let transactions = self.clear_given_keys(debtor_keys, creditor_keys);
         self.panic_unless_empty();
-        return transactions;
+        transactions
     }
 
-    // Settles a specified list of debtors and creditors against in other, in random order.
+    // Settles a specified list of debtors and creditors, in random order.
     fn clear_given_keys(
         &mut self,
         debtors: Vec<String>,
@@ -288,44 +294,50 @@ impl Ledger {
         let mut payments: Vec<Transaction> = Vec::new();
 
         for debtor in &debtors {
-            let mut debtor_amount = *self.map.get(debtor).unwrap();
+            let mut debtor_amount = self.map.get(debtor).unwrap().clone();
 
             for creditor in &creditors {
-                let mut creditor_amount = *self.map.get(creditor).unwrap();
+                let mut creditor_amount = self.map.get(creditor).unwrap().clone();
 
                 // If there's still debt and credit, create a payment.
                 // If either one is missing, try grabbing another creditor
                 // If you run out of creditors, grab another debtor and start again.
-                while (creditor_amount > 0) && (debtor_amount < 0) {
-                    let credit_abs = creditor_amount.abs();
-                    let debt_abs = debtor_amount.abs();
+                while (creditor_amount.is_positive()) && (debtor_amount.is_negative()) {
+                    let credit_abs = creditor_amount.amount().abs();
+                    let debt_abs = debtor_amount.amount().abs();
                     let payment_amount = cmp::min(credit_abs, debt_abs);
 
-                    debtor_amount += payment_amount;
-                    self.map.insert(debtor.clone(), debtor_amount);
+                    debtor_amount += Money::new(payment_amount, Currency::new("USD".to_string()));
+                    self.map.insert(debtor.clone(), debtor_amount.clone());
 
-                    creditor_amount -= payment_amount;
-                    self.map.insert(creditor.clone(), creditor_amount);
+                    creditor_amount -= Money::new(payment_amount, Currency::new("USD".to_string()));
+                    self.map.insert(creditor.clone(), creditor_amount.clone());
 
-                    payments.push(
-                        Transaction::new(debtor.clone(), creditor.clone(), payment_amount).unwrap(),
-                    )
+                    payments.push(transaction!(
+                        debtor.clone(),
+                        creditor.clone(),
+                        Money::new(payment_amount, Currency::new("USD".to_string()))
+                    ));
                 }
             }
         }
-        return payments;
+        payments
     }
 
     // Finds zero sum combinations of a given size of ledger entries.
-    fn find_zero_sum_combinations(&self, combo_size: usize) -> Vec<Vec<(String, i32)>> {
-        let mut zero_sum_combinations: Vec<Vec<(String, i32)>> = Vec::new();
+    fn find_zero_sum_combinations(&self, combo_size: usize) -> Vec<Vec<(String, Money)>> {
+        let mut zero_sum_combinations: Vec<Vec<(String, Money)>> = Vec::new();
         let combinations = self.to_vector().into_iter().combinations(combo_size);
         for item in combinations {
-            if item.iter().fold(0, |acc, x| acc + x.1) == 0 {
+            if item
+                .iter()
+                .fold(money!(0, "USD"), |acc, x| acc + x.1.clone())
+                .is_zero()
+            {
                 zero_sum_combinations.push(item);
             }
         }
-        return zero_sum_combinations;
+        zero_sum_combinations
     }
 
     // Returns vectors of keys of debtors and creditors with an active balance.s
@@ -334,14 +346,14 @@ impl Ledger {
         let mut debtors: Vec<String> = Vec::new();
 
         for (person, value) in &self.map {
-            if *value > 0 {
+            if value.is_positive() {
                 creditors.push(person.clone());
-            } else if *value < 0 {
+            } else if value.is_negative() {
                 debtors.push(person.clone());
             } else {
             }
         }
-        return (debtors, creditors);
+        (debtors, creditors)
     }
 }
 
@@ -353,24 +365,23 @@ mod tests {
     // This allows two entries in the ledger to be removed in exchange for a single payment.
     // For example, if A = -10 and B = +10, they should always first over any other possibility
     #[test]
-    fn matches_equal_debts_and_credits_when_groupsize_is_2() {
+    fn ledger_settle_matches_equal_debts_and_credits() {
         let mut ledger = Ledger::new();
 
         let expected_results = vec![
-            Transaction::new("A".to_string(), "B".to_string(), 2).unwrap(),
-            Transaction::new("C".to_string(), "F".to_string(), 3).unwrap(),
-            Transaction::new("D".to_string(), "F".to_string(), 5).unwrap(),
-            Transaction::new("E".to_string(), "F".to_string(), 7).unwrap(),
+            transaction!("A", "B", money!(2, "USD")),
+            transaction!("C", "F", money!(3, "USD")),
+            transaction!("D", "F", money!(5, "USD")),
+            transaction!("E", "F", money!(7, "USD")),
         ];
 
         // The worst case match (i.e. random) can accidentially find the optimal solution for small
         // sets, so we repeat to make this very unlikely
         for _ in 0..5 {
-            ledger.add_transaction(Transaction::new("A".to_string(), "B".to_string(), 2).unwrap());
-            ledger.add_transaction(Transaction::new("C".to_string(), "F".to_string(), 3).unwrap());
-            ledger.add_transaction(Transaction::new("D".to_string(), "F".to_string(), 5).unwrap());
-            ledger.add_transaction(Transaction::new("E".to_string(), "F".to_string(), 7).unwrap());
-
+            ledger.add_transaction(transaction!("A", "B", money!(2, "USD")));
+            ledger.add_transaction(transaction!("C", "F", money!(3, "USD")));
+            ledger.add_transaction(transaction!("D", "F", money!(5, "USD")));
+            ledger.add_transaction(transaction!("E", "F", money!(7, "USD")));
             let mut payments = ledger.settle(2);
             payments.sort();
             assert_eq!(payments, expected_results);
@@ -381,28 +392,28 @@ mod tests {
     // This allows three entries in the ledger to be removed in exchange for two payments.
     // For example, if A = -10,  B = +5, C= +5.
     #[test]
-    fn finds_groups_of_three_credits_and_debits_when_groupsize_is_3() {
+    fn ledger_settle_with_size_3_matches_groups_of_3_credits_and_debits() {
         // Test that group matched  payments are always settled first.
         let mut ledger = Ledger::new();
 
         let expected_results = vec![
-            Transaction::new("A".to_string(), "D".to_string(), 3).unwrap(),
-            Transaction::new("C".to_string(), "D".to_string(), 4).unwrap(),
-            Transaction::new("E".to_string(), "B".to_string(), 10).unwrap(),
-            Transaction::new("F".to_string(), "B".to_string(), 17).unwrap(),
-            Transaction::new("J".to_string(), "K".to_string(), 20).unwrap(),
-            Transaction::new("U".to_string(), "K".to_string(), 21).unwrap(),
+            transaction!("A", "D", money!(3, "USD")),
+            transaction!("C", "D", money!(4, "USD")),
+            transaction!("E", "B", money!(10, "USD")),
+            transaction!("F", "B", money!(17, "USD")),
+            transaction!("J", "K", money!(20, "USD")),
+            transaction!("U", "K", money!(21, "USD")),
         ];
 
         // The worst case match (i.e. random) can accidentially find the optimal solution for small
         // sets, so we repeat to make this very unlikely
         for _ in 0..5 {
-            ledger.add_transaction(Transaction::new("A".to_string(), "D".to_string(), 3).unwrap());
-            ledger.add_transaction(Transaction::new("C".to_string(), "D".to_string(), 4).unwrap());
-            ledger.add_transaction(Transaction::new("E".to_string(), "B".to_string(), 10).unwrap());
-            ledger.add_transaction(Transaction::new("F".to_string(), "B".to_string(), 17).unwrap());
-            ledger.add_transaction(Transaction::new("J".to_string(), "K".to_string(), 20).unwrap());
-            ledger.add_transaction(Transaction::new("U".to_string(), "K".to_string(), 21).unwrap());
+            ledger.add_transaction(transaction!("A", "D", money!(3, "USD")));
+            ledger.add_transaction(transaction!("C", "D", money!(4, "USD")));
+            ledger.add_transaction(transaction!("E", "B", money!(10, "USD")));
+            ledger.add_transaction(transaction!("F", "B", money!(17, "USD")));
+            ledger.add_transaction(transaction!("J", "K", money!(20, "USD")));
+            ledger.add_transaction(transaction!("U", "K", money!(21, "USD")));
 
             let mut payments = ledger.settle(3);
             payments.sort();
@@ -412,63 +423,75 @@ mod tests {
 
     #[test]
     #[should_panic]
-    fn panics_when_settling_unbalanced_ledger() {
+    fn ledger_settle_panics_if_unbalanced() {
         let mut ledger = Ledger::new();
-        ledger.map.entry("A".to_string()).or_insert(10);
+        ledger
+            .map
+            .entry("A".to_string())
+            .or_insert(money!(10, "USD"));
         ledger.settle(2);
     }
 
-    // Multi Party Transaction Tests
+    //
+    // Multi-Party Transaction Tests
+    //
+
     #[test]
-    fn can_handle_debtor_rounding() {
+    fn mptx_can_handle_debtor_rounding() {
         let transaction = MultiPartyTransaction::new(
             vec!["A".to_string(), "B".to_string(), "C".to_string()],
             vec!["D".to_string()],
-            10,
+            money!(10, "USD"),
         )
         .unwrap();
         let mut ledger = Ledger::new();
         ledger.add_multi_party_transaction(transaction);
-        let ledger_balance = ledger.to_vector().into_iter().fold(0, |acc, x| acc + x.1);
-        assert_eq!(ledger_balance, 0);
+        let remaining = ledger
+            .to_vector()
+            .into_iter()
+            .fold(money!(0, "USD"), |acc, x| acc + x.1);
+        assert_eq!(remaining, money!(0, "USD"));
     }
 
     #[test]
-    fn can_handle_creditor_rounding() {
+    fn mptx_can_handle_creditor_rounding() {
         let transaction = MultiPartyTransaction::new(
             vec!["A".to_string()],
             vec!["B".to_string(), "C".to_string(), "D".to_string()],
-            10,
+            money!(10, "USD"),
         )
         .unwrap();
         let mut ledger = Ledger::new();
         ledger.add_multi_party_transaction(transaction);
-        let ledger_balance = ledger.to_vector().into_iter().fold(0, |acc, x| acc + x.1);
-        assert_eq!(ledger_balance, 0);
+        let ledger_balance = ledger
+            .to_vector()
+            .into_iter()
+            .fold(money!(0, "USD"), |acc, x| acc + x.1);
+        assert_eq!(ledger_balance, money!(0, "USD"));
     }
 
+    //
     // Transaction Tests
+    //
+
     #[test]
-    fn can_create_positive_transaction() {
-        match Transaction::new("A".to_string(), "B".to_string(), 1) {
+    fn tx_can_create_positive_transaction() {
+        match Transaction::new("A".to_string(), "B".to_string(), money!(1, "USD")) {
             Ok(_) => assert!(true),
             Err(_) => assert!(false),
         };
     }
 
     #[test]
-    fn cannot_create_negative_transaction() {
-        match Transaction::new("A".to_string(), "B".to_string(), -1) {
+    fn tx_cannot_create_non_positive_transaction() {
+        match Transaction::new("A".to_string(), "B".to_string(), money!(-1, "USD")) {
             Ok(_) => assert!(false),
-            Err(_) => assert!(true),
+            Err(_) => assert!(true), // TODO: catch the right error here
         };
-    }
 
-    #[test]
-    fn cannot_create_zero_transaction() {
-        match Transaction::new("A".to_string(), "B".to_string(), -1) {
+        match Transaction::new("A".to_string(), "B".to_string(), money!(0, "USD")) {
             Ok(_) => assert!(false),
-            Err(_) => assert!(true),
+            Err(_) => assert!(true), // TODO: catch the right error here
         };
     }
 }