+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 * Transaction API: Multiparty debts in 1 transaction. (e.g. Alice & Bob paid 50 for a dinner shared between Alice, Bob, Charlie and Danielle)
+* Use a rich 'Money' Type that handles rounding and currency metadata. 
 
 ## [0.1.0] - 2019-11-19
 
@@ -11,7 +12,6 @@
 
 ### v0.2.0
 * Settle API: Auto-detect best group-size, and provide API for manual sizing.
-* Use a rich 'Money' Type that handles rounding and currency metadata. 
 
 ### v0.3.0
 * Implement a true subset sum solution: https://pure.tue.nl/ws/files/2062204/623903.pdf